Modified the hyperlinks for signup and login
The hyperlink for the signup is added and both the signup and the login hyperlinks are now made to take user into specific parts of the login webpage where if login is clicked, user is taken to login form in the login.html file. While if the user clicks on sign up, user is taken to the sign up form instead within the login.html file!